test and fix cap()
  - make unity builds work, test internals with them
  - swap from assert() to expect()

+++ buffer.c
@@ -4,7 +4,7 @@
 
 static size_t cap(size_t len) {
     return len <= 1 ? len
-                    : 1 << (CHAR_BIT*sizeof(len) - (size_t)__builtin_clzll(len-1));
+                    : (size_t)1 << (CHAR_BIT*sizeof(len) - (size_t)__builtin_clzll(len-1));
 }
 
 struct buffer* buffer_grow(struct buffer *b, size_t more) {
